Refactoring ColissimoRule for relay pickup (46736) (#349)
* refs #46736 ColissimoRule

* refs #46736 removed CdiscountColissimo

* refs #46736 tests

* refs #46736 tests
