+++ 202/tests/OrderImport/dataset/order-zalando.json
@@ -18,7 +18,7 @@
        "method":"" 
     },
     "shipment":{
-       "carrier":"",
+       "carrier":"Colissimo",
        "trackingNumber":"" 
     },
     "additionalFields":{
